Finished attack animation.

+++ Scrolls of the Elders/Scrolls of the Elders/Player.h
@@ -42,7 +42,7 @@ public:
 	string scoreString;
 	float currentAttackAngle;
 	float startingAngle;
-	int stage;
+	float stage;
 	bool attackState;
 	bool attacked;
